klient: Moved Machine struct to it's own package
