+++ go/src/koding/klient/remote/kitepinger/kitepinger_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/koding/klient/Godeps/_workspace/src/github.com/koding/kite"
+	"github.com/koding/kite"
 )
 
 func TestRegister(t *testing.T) {
