+++ go/src/koding/klient/remote/mountfolder_test.go
@@ -6,7 +6,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/koding/klient/Godeps/_workspace/src/github.com/koding/kite"
+	"github.com/koding/kite"
 )
 
 func TestCheckSizeOfRemotefolder(t *testing.T) {
